refactor main merge working
