+++ src/data.cc
@@ -4,6 +4,7 @@ license: GNU/GPL v3
 */
 
 #include "data.h"
+#include "rnd.h"
 
 //#include "node/node.h"
 //external includes
