+++ src/node/nodeadd.h
@@ -20,20 +20,7 @@ namespace FT{
             /// Evaluates the node symbolically
             void eval_eqn(Stacks& stack);
 
-            ArrayXd getDerivative(Trace& stack, int loc) 
-            {
-                switch (loc) {
-                    case 3: // d/dW[1] 
-                        return stack.f[stack.f.size()-2];
-                    case 2: // d/dW[0]
-                        return stack.f[stack.f.size()-1];
-                    case 1: // d/dx2
-                        return this->W[1] * ArrayXd::Ones(stack.f[stack.f.size()-2].size());
-                    case 0: // d/dx1
-                    default:
-                        return this->W[0] * ArrayXd::Ones(stack.f[stack.f.size()-1].size());
-                } 
-            }
+            ArrayXd getDerivative(Trace& stack, int loc);
 
         protected:
             NodeAdd* clone_impl() const override;
